Add an algorithm to sniff MP3 without ID3
This works by trying to successfully match two mp3 capture pattern, one at the
beginning of the stream, and one at an offset in the byte stream that we can
determine by parsing the header of the first mp3 packet.

To implement this, we need to bump the maximum size to sniff to 1445, which is
the maximum byte offset at which we can find the mp3 capture pattern in the
second mp3 packet in an mp3-without-id3 byte stream.

Fixes https://www.w3.org/Bugs/Public/show_bug.cgi?id=24220.

PR: https://github.com/whatwg/mimesniff/pull/4