Use a more common serialization for MIME types
This is a defense-in-depth for the issues that arised at https://github.com/whatwg/mimesniff/issues/84. In theory they are all solved by https://github.com/whatwg/xhr/pull/224, but given how quickly several sites turned up that relied on a space, it seems prudent to standardize on that serialization to avoid similar issues elsewhere.